Client, ServerGUI and ClientGUI
ServerGUI is a simple GUI using JTextArea don't expect fancy fonts,
colors, Icons,... I kept it as simple as possible.
The ClientGUI class.
ClientGUI is for the Client. Also a BorderLayout. In the NORTH region
two JTextField to enter the host name of the Server and the port number
it is listening to.